+++ project/python/{{ cookiecutter.project_name }}/dags/{{ cookiecutter.project_name }}.py
@@ -28,6 +28,7 @@ DatafyContainerOperator(
     name="sample",
     image=image,
     arguments=["{% raw %}--date", "{{ ds }}", "--jobs", "sample", "--env", "{{ macros.env() }}{% endraw %}"],
+    instance_type="mx_micro",
 {%- if cookiecutter.role_creation != "none" %}
     service_account_name="{{ cookiecutter.project_name }}"
 {%- endif %}